refactor(blocks): improve the configuration of more menu (#8089)
* caches more menu configuration
* adds `getMoreMenuConfig` method
* splits context and config files

BTW: currently, this can only be configured in `firstUpdated`, because it depends on `std`
