+++ packages/blocks/src/root-block/widgets/code-toolbar/components/code-toolbar.ts
@@ -13,7 +13,7 @@ import { LitElement, css, html } from 'lit';
 import { customElement, property, query, state } from 'lit/decorators.js';
 import { styleMap } from 'lit/directives/style-map.js';
 
-import type { CodeBlockToolbarContext } from '../type.js';
+import type { CodeBlockToolbarContext } from '../context.js';
 
 @customElement('affine-code-toolbar')
 export class AffineCodeToolbar extends WithDisposable(LitElement) {
@@ -73,8 +73,6 @@ export class AffineCodeToolbar extends WithDisposable(LitElement) {
 
     assertExists(this._moreButton);
 
-    const groups = this.context.config.configure(this.moreGroups);
-
     createLitPortal({
       template: html`
         <editor-menu-content
@@ -86,7 +84,7 @@ export class AffineCodeToolbar extends WithDisposable(LitElement) {
           })}
         >
           <div data-size="large" data-orientation="vertical">
-            ${renderGroups(groups, this.context)}
+            ${renderGroups(this.moreGroups, this.context)}
           </div>
         </editor-menu-content>
       `,
