+++ packages/blocks/src/root-block/widgets/element-toolbar/more-menu/config.ts
@@ -383,3 +383,12 @@ export const deleteGroup: MenuItemGroup<ElementToolbarMoreMenuContext> = {
     },
   ],
 };
+
+export const BUILT_IN_GROUPS = [
+  sectionGroup,
+  reorderGroup,
+  openGroup,
+  clipboardGroup,
+  conversionsGroup,
+  deleteGroup,
+];
