+++ packages/blocks/src/root-block/widgets/embed-card-toolbar/context.ts
@@ -1,0 +1,45 @@
+import type { EmbedToolbarBlockComponent } from '../../../_common/components/embed-card/type.js';
+
+import { MenuContext } from '../../configs/toolbar.js';
+
+export class EmbedCardToolbarContext extends MenuContext {
+  override close = () => {
+    this.abortController.abort();
+  };
+
+  constructor(
+    public blockComponent: EmbedToolbarBlockComponent,
+    public abortController: AbortController
+  ) {
+    super();
+  }
+
+  isEmpty() {
+    return false;
+  }
+
+  isMultiple() {
+    return false;
+  }
+
+  isSingle() {
+    return true;
+  }
+
+  get doc() {
+    return this.blockComponent.doc;
+  }
+
+  get host() {
+    return this.blockComponent.host;
+  }
+
+  get selectedBlockModels() {
+    if (this.blockComponent.model) return [this.blockComponent.model];
+    return [];
+  }
+
+  get std() {
+    return this.host.std;
+  }
+}
