+++ packages/blocks/src/root-block/widgets/image-toolbar/context.ts
