+++ packages/blocks/src/root-block/widgets/surface-ref-toolbar/surface-ref-toolbar.ts
@@ -4,16 +4,12 @@ import { HoverController } from '@blocksuite/affine-components/hover';
 import {
   CaptionIcon,
   CenterPeekIcon,
-  CopyIcon,
-  DeleteIcon,
-  DownloadIcon,
   EdgelessModeIcon,
   MoreVerticalIcon,
   OpenIcon,
   SmallArrowDownIcon,
 } from '@blocksuite/affine-components/icons';
 import { isPeekable, peek } from '@blocksuite/affine-components/peek';
-import { toast } from '@blocksuite/affine-components/toast';
 import {
   type MenuItem,
   type MenuItemGroup,
@@ -21,7 +17,6 @@ import {
   renderGroups,
   renderToolbarSeparator,
 } from '@blocksuite/affine-components/toolbar';
-import { downloadBlob } from '@blocksuite/affine-shared/utils';
 import { WidgetComponent } from '@blocksuite/block-std';
 import { offset, shift } from '@floating-ui/dom';
 import { html, nothing } from 'lit';
@@ -31,11 +26,11 @@ import { join } from 'lit/directives/join.js';
 import { repeat } from 'lit/directives/repeat.js';
 
 import type { SurfaceRefBlockComponent } from '../../../surface-ref-block/index.js';
-import type { EdgelessRootPreviewBlockComponent } from '../../edgeless/edgeless-root-preview-block.js';
 
 import { PAGE_HEADER_HEIGHT } from '../../../_common/consts.js';
-import { MenuContext } from '../../configs/toolbar.js';
-import { edgelessToBlob, writeImageBlobToClipboard } from './utils.js';
+import { getMoreMenuConfig } from '../../configs/toolbar.js';
+import { BUILT_IN_GROUPS } from './config.js';
+import { SurfaceRefToolbarContext } from './context.js';
 
 export const AFFINE_SURFACE_REF_TOOLBAR = 'affine-surface-ref-toolbar';
 
@@ -69,9 +64,8 @@ export class AffineSurfaceRefToolbar extends WidgetComponent<
 
       return {
         template: SurfaceRefToolbarOptions({
-          block: this.block,
-          model: this.block.model,
-          abortController,
+          context: new SurfaceRefToolbarContext(this.block, abortController),
+          groups: this.moreGroups,
         }),
         computePosition: {
           referenceElement: this.block,
@@ -96,9 +90,17 @@ export class AffineSurfaceRefToolbar extends WidgetComponent<
     }
   );
 
+  /*
+   * Caches the more menu items.
+   * Currently only supports configuring more menu.
+   */
+  moreGroups: MenuItemGroup<SurfaceRefToolbarContext>[] =
+    cloneGroups(BUILT_IN_GROUPS);
+
   override connectedCallback() {
     super.connectedCallback();
 
+    this.moreGroups = getMoreMenuConfig(this.std).configure(this.moreGroups);
     this._hoverController.setReference(this.block);
   }
 }
@@ -109,130 +111,16 @@ declare global {
   }
 }
 
-export class SurfaceRefToolbarContext extends MenuContext {
-  constructor(
-    public blockComponent: SurfaceRefBlockComponent,
-    public abortController: AbortController
-  ) {
-    super();
-  }
-
-  isEmpty() {
-    return !this.blockComponent;
-  }
-
-  isMultiple() {
-    return false;
-  }
-
-  isSingle() {
-    return true;
-  }
-
-  get doc() {
-    return this.blockComponent.doc;
-  }
-
-  get host() {
-    return this.blockComponent.host;
-  }
-
-  get selectedBlockModels() {
-    if (this.blockComponent) return [this.blockComponent.model];
-    return [];
-  }
-
-  get std() {
-    return this.host.std;
-  }
-}
-
-const BUILT_IN_GROUPS: MenuItemGroup<SurfaceRefToolbarContext>[] = [
-  {
-    type: 'clipboard',
-    when: ctx => !!(ctx.blockComponent.referenceModel && ctx.doc.root),
-    items: [
-      {
-        type: 'copy',
-        label: 'Copy',
-        icon: CopyIcon,
-        action: ctx => {
-          if (!(ctx.blockComponent.referenceModel && ctx.doc.root?.id)) return;
-
-          const referencedModel = ctx.blockComponent.referenceModel;
-          const editor = ctx.blockComponent.previewEditor;
-          const edgelessRootElement = editor?.view.getBlock(ctx.doc.root.id);
-          const surfaceRenderer = (
-            edgelessRootElement as EdgelessRootPreviewBlockComponent
-          )?.surface?.renderer;
-
-          edgelessToBlob(ctx.host, {
-            surfaceRefBlock: ctx.blockComponent,
-            surfaceRenderer,
-            edgelessElement: referencedModel,
-          })
-            .then(blob => writeImageBlobToClipboard(blob))
-            .then(() => toast(ctx.host, 'Copied image to clipboard'))
-            .catch(console.error);
-        },
-      },
-      {
-        type: 'download',
-        label: 'Download',
-        icon: DownloadIcon,
-        action: ctx => {
-          if (!(ctx.blockComponent.referenceModel && ctx.doc.root?.id)) return;
-
-          const referencedModel = ctx.blockComponent.referenceModel;
-          const editor = ctx.blockComponent.previewEditor;
-          const edgelessRootElement = editor?.view.getBlock(ctx.doc.root.id);
-          const surfaceRenderer = (
-            edgelessRootElement as EdgelessRootPreviewBlockComponent
-          )?.surface?.renderer;
-
-          edgelessToBlob(ctx.host, {
-            surfaceRefBlock: ctx.blockComponent,
-            surfaceRenderer,
-            edgelessElement: referencedModel,
-          })
-            .then(blob => {
-              const fileName =
-                'title' in referencedModel
-                  ? (referencedModel.title?.toString() ?? 'Edgeless Content')
-                  : 'Edgeless Content';
-
-              downloadBlob(blob, fileName);
-            })
-            .catch(console.error);
-        },
-      },
-    ],
-  },
-  {
-    type: 'delete',
-    items: [
-      {
-        type: 'delete',
-        label: 'Delete',
-        icon: DeleteIcon,
-        disabled: ctx => ctx.doc.readonly,
-        action: ctx => {
-          ctx.doc.deleteBlock(ctx.blockComponent.model);
-          ctx.abortController.abort();
-        },
-      },
-    ],
-  },
-];
-
-function SurfaceRefToolbarOptions(options: {
-  block: SurfaceRefBlockComponent;
-  model: SurfaceRefBlockModel;
-  abortController: AbortController;
+function SurfaceRefToolbarOptions({
+  context,
+  groups,
+}: {
+  context: SurfaceRefToolbarContext;
+  groups: MenuItemGroup<SurfaceRefToolbarContext>[];
 }) {
-  const { block, model, abortController } = options;
-  const readonly = model.doc.readonly;
-  const hasValidReference = !!block.referenceModel;
+  const { blockComponent, abortController } = context;
+  const readonly = blockComponent.model.doc.readonly;
+  const hasValidReference = !!blockComponent.referenceModel;
 
   const openMenuActions: MenuItem[] = [];
   if (hasValidReference) {
@@ -240,22 +128,20 @@ function SurfaceRefToolbarOptions(options: {
       type: 'open-in-edgeless',
       label: 'Open in edgeless',
       icon: EdgelessModeIcon,
-      action: () => block.viewInEdgeless(),
+      action: () => blockComponent.viewInEdgeless(),
       disabled: readonly,
     });
 
-    if (isPeekable(block)) {
+    if (isPeekable(blockComponent)) {
       openMenuActions.push({
         type: 'open-in-center-peek',
         label: 'Open in center peek',
         icon: CenterPeekIcon,
-        action: () => peek(block),
+        action: () => peek(blockComponent),
       });
     }
   }
 
-  const context = new SurfaceRefToolbarContext(block, abortController);
-  const groups = context.config.configure(cloneGroups(BUILT_IN_GROUPS));
   const moreMenuActions = renderGroups(groups, context);
 
   const buttons = [
@@ -299,7 +185,7 @@ function SurfaceRefToolbarOptions(options: {
             aria-label="Caption"
             @click=${() => {
               abortController.abort();
-              block.captionElement.show();
+              blockComponent.captionElement.show();
             }}
           >
             ${CaptionIcon}
